2/7 commit

+++ lib/ui/views/archived_customer/archived_customer_viewmodel.dart
@@ -4,12 +4,15 @@ import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 import 'package:verzo/app/app.dialogs.dart';
 import 'package:verzo/app/app.locator.dart';
+import 'package:verzo/app/app.router.dart';
+import 'package:verzo/services/authentication_service.dart';
 import 'package:verzo/services/sales_service.dart';
 import 'package:verzo/ui/common/database_helper.dart';
 
 class ArchivedCustomerViewModel extends FutureViewModel<List<Customers>> {
   final navigationService = locator<NavigationService>();
   final _saleService = locator<SalesService>();
+  final authService = locator<AuthenticationService>();
   final DialogService dialogService = locator<DialogService>();
 
   List<Customers> customers = [];
@@ -19,10 +22,14 @@ class ArchivedCustomerViewModel extends FutureViewModel<List<Customers>> {
   Future<List<Customers>> getArchivedCustomersByBusiness() async {
     SharedPreferences prefs = await SharedPreferences.getInstance();
     String businessIdValue = prefs.getString('businessId') ?? '';
-
-    // Retrieve existing customers
-    customers = await _saleService.getArchivedCustomerByBusiness(
-        businessId: businessIdValue);
+    final result = await authService.refreshToken();
+    if (result.error != null) {
+      await navigationService.replaceWithLoginView();
+    } else if (result.tokens != null) {
+      // Retrieve existing customers
+      customers = await _saleService.getArchivedCustomerByBusiness(
+          businessId: businessIdValue);
+    }
 
     rebuildUi();
 
@@ -43,27 +50,33 @@ class ArchivedCustomerViewModel extends FutureViewModel<List<Customers>> {
 
     // Check if the user confirmed the action
     if (response?.confirmed == true) {
-      // Proceed with archiving if confirmed
-      final bool isUnArchived =
-          await _saleService.unArchiveCustomer(customerId: customerId);
+      final result = await authService.refreshToken();
+      if (result.error != null) {
+        await navigationService.replaceWithLoginView();
+      } else if (result.tokens != null) {
+        // Proceed with archiving if confirmed
+        final bool isUnArchived =
+            await _saleService.unArchiveCustomer(customerId: customerId);
 
-      if (isUnArchived) {
-        await dialogService.showCustomDialog(
-            variant: DialogType.archiveSuccess,
-            title: 'Unarchived!',
-            description: 'Your customer has been successfully unarchived.',
-            barrierDismissible: true,
-            mainButtonTitle: 'Ok');
-        await db.delete('customers');
+        if (isUnArchived) {
+          await dialogService.showCustomDialog(
+              variant: DialogType.archiveSuccess,
+              title: 'Unarchived!',
+              description: 'Your customer has been successfully unarchived.',
+              barrierDismissible: true,
+              mainButtonTitle: 'Ok');
+          await db.delete('customers');
+        }
+
+        reloadCustomer();
+
+        return isUnArchived;
       }
-
-      reloadCustomer();
-
-      return isUnArchived;
     } else {
       // User canceled the action
       return false;
     }
+    return false;
   }
 
   Future<bool> deleteCustomer(String customerId) async {
@@ -80,28 +93,34 @@ class ArchivedCustomerViewModel extends FutureViewModel<List<Customers>> {
 
     // Check if the user confirmed the action
     if (response?.confirmed == true) {
-      // Proceed with deleting if confirmed
-      final bool isDeleted =
-          await _saleService.deleteCustomer(customerId: customerId);
+      final result = await authService.refreshToken();
+      if (result.error != null) {
+        await navigationService.replaceWithLoginView();
+      } else if (result.tokens != null) {
+        // Proceed with deleting if confirmed
+        final bool isDeleted =
+            await _saleService.deleteCustomer(customerId: customerId);
 
-      if (isDeleted) {
-        await dialogService.showCustomDialog(
-            variant: DialogType.deleteSuccess,
-            title: 'Deleted!',
-            description: 'Your customer has been successfully deleted.',
-            barrierDismissible: true,
-            mainButtonTitle: 'Ok');
-        await db.delete('customers');
+        if (isDeleted) {
+          await dialogService.showCustomDialog(
+              variant: DialogType.deleteSuccess,
+              title: 'Deleted!',
+              description: 'Your customer has been successfully deleted.',
+              barrierDismissible: true,
+              mainButtonTitle: 'Ok');
+          await db.delete('customers');
+        }
+
+        // Navigate to the customer view
+        reloadCustomer();
+
+        return isDeleted;
       }
-
-      // Navigate to the customer view
-      reloadCustomer();
-
-      return isDeleted;
     } else {
       // User canceled the action
       return false;
     }
+    return false;
   }
 
   void reloadCustomer() async {
