+++ lib/ui/views/archived_expense/archived_expense_viewmodel.dart
@@ -4,12 +4,15 @@ import 'package:stacked/stacked.dart';
 import 'package:stacked_services/stacked_services.dart';
 import 'package:verzo/app/app.dialogs.dart';
 import 'package:verzo/app/app.locator.dart';
+import 'package:verzo/app/app.router.dart';
+import 'package:verzo/services/authentication_service.dart';
 import 'package:verzo/services/expense_service.dart';
 import 'package:verzo/ui/common/database_helper.dart';
 
 class ArchivedExpenseViewModel extends FutureViewModel<List<Expenses>> {
   final navigationService = locator<NavigationService>();
   final _expenseService = locator<ExpenseService>();
+  final authService = locator<AuthenticationService>();
   final DialogService dialogService = locator<DialogService>();
 
   List<Expenses> expenses = [];
@@ -20,10 +23,13 @@ class ArchivedExpenseViewModel extends FutureViewModel<List<Expenses>> {
   Future<List<Expenses>> getArchivedExpenseByBusiness() async {
     SharedPreferences prefs = await SharedPreferences.getInstance();
     String businessIdValue = prefs.getString('businessId') ?? '';
-
-    // Retrieve existing expense categories
-    expenses = await _expenseService.getArchivedExpenseByBusiness(
-        businessId: businessIdValue);
+    final result = await authService.refreshToken();
+    if (result.error != null) {
+      await navigationService.replaceWithLoginView();
+    } else if (result.tokens != null) {
+      expenses = await _expenseService.getArchivedExpenseByBusiness(
+          businessId: businessIdValue);
+    }
 
     rebuildUi();
     return expenses;
@@ -45,30 +51,36 @@ class ArchivedExpenseViewModel extends FutureViewModel<List<Expenses>> {
         // confirmationTitle: 'Ok',
         );
     if (response?.confirmed == true) {
-      final bool isDeleted =
-          await _expenseService.deleteExpense(expenseId: expenseId);
+      final result = await authService.refreshToken();
+      if (result.error != null) {
+        await navigationService.replaceWithLoginView();
+      } else if (result.tokens != null) {
+        final bool isDeleted =
+            await _expenseService.deleteExpense(expenseId: expenseId);
 
-      if (isDeleted) {
-        // await futureToRun();
-        await dialogService.showCustomDialog(
-          variant: DialogType.deleteSuccess,
-          title: 'Deleted!',
-          description: 'Your expense has been successfully deleted.',
-          barrierDismissible: true,
-          mainButtonTitle: 'Ok',
-        );
-        await db.delete('expenses');
-        await db2.delete('expenses');
-        await dbExpenseWeek.delete('expenses_for_week');
-        await dbExpenseMonth.delete('expenses_for_month');
+        if (isDeleted) {
+          // await futureToRun();
+          await dialogService.showCustomDialog(
+            variant: DialogType.deleteSuccess,
+            title: 'Deleted!',
+            description: 'Your expense has been successfully deleted.',
+            barrierDismissible: true,
+            mainButtonTitle: 'Ok',
+          );
+          await db.delete('expenses');
+          await db2.delete('expenses');
+          await dbExpenseWeek.delete('expenses_for_week');
+          await dbExpenseMonth.delete('expenses_for_month');
+        }
+
+        reloadExpense();
+        return isDeleted;
       }
-
-      reloadExpense();
-      return isDeleted;
     } else {
       // User canceled the action
       return false;
     }
+    return false;
   }
 
   Future<bool> unArchiveExpense(String expenseId) async {
@@ -89,28 +101,34 @@ class ArchivedExpenseViewModel extends FutureViewModel<List<Expenses>> {
         );
 
     if (response?.confirmed == true) {
-      final bool isUnArchived =
-          await _expenseService.unarchiveExpense(expenseId: expenseId);
+      final result = await authService.refreshToken();
+      if (result.error != null) {
+        await navigationService.replaceWithLoginView();
+      } else if (result.tokens != null) {
+        final bool isUnArchived =
+            await _expenseService.unarchiveExpense(expenseId: expenseId);
 
-      if (isUnArchived) {
-        await dialogService.showCustomDialog(
-            variant: DialogType.archiveSuccess,
-            title: 'Unarchived!',
-            description: 'Your expense has been successfully unarchived.',
-            barrierDismissible: true,
-            mainButtonTitle: 'Ok');
+        if (isUnArchived) {
+          await dialogService.showCustomDialog(
+              variant: DialogType.archiveSuccess,
+              title: 'Unarchived!',
+              description: 'Your expense has been successfully unarchived.',
+              barrierDismissible: true,
+              mainButtonTitle: 'Ok');
 
-        await db.delete('expenses');
-        await db2.delete('expenses');
-        await dbExpenseWeek.delete('expenses_for_week');
-        await dbExpenseMonth.delete('expenses_for_month');
+          await db.delete('expenses');
+          await db2.delete('expenses');
+          await dbExpenseWeek.delete('expenses_for_week');
+          await dbExpenseMonth.delete('expenses_for_month');
+        }
+        reloadExpense();
+        return isUnArchived;
       }
-      reloadExpense();
-      return isUnArchived;
     } else {
       // User canceled the action
       return false;
     }
+    return false;
   }
 
   void reloadExpense() async {
